backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
celery[redis]==5.4.0
redis==5.1.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
weasyprint==62.3
jinja2==3.1.4
python-dateutil==2.9.0
dnspython==2.6.1
ipaddress==1.0.23
geoip2==4.8.0
